# feat: use pools for attributes and options