--- a/operation_name_go124.go
+++ b/operation_name_go124.go
@@ -7,13 +7,9 @@ import (
 	"strings"
 )
 
-// sqlOperationName attempts to get the first 'word' from a given SQL query, which usually
+// defaultSpanNameFunc attempts to get the first 'word' from a given SQL query, which usually
 // is the operation name (e.g. 'SELECT').
-func (t *Tracer) sqlOperationName(stmt string) string {
-	if t.spanNameFunc != nil {
-		return t.spanNameFunc(stmt)
-	}
-
+func defaultSpanNameFunc(stmt string) string {
 	for word := range strings.FieldsSeq(stmt) {
 		return strings.ToUpper(word)
 	}