--- a/operation_name_legacy.go
+++ b/operation_name_legacy.go
@@ -8,16 +8,9 @@ import (
 	"unicode"
 )
 
-// sqlOperationName attempts to get the first 'word' from a given SQL query, which usually
+// defaultSpanNameFunc attempts to get the first 'word' from a given SQL query, which usually
 // is the operation name (e.g. 'SELECT').
-func (t *Tracer) sqlOperationName(stmt string) string {
-	// If a custom function is provided, use that. Otherwise, fall back to the
-	// default implementation. This allows users to override the default
-	// behavior without having to reimplement it.
-	if t.spanNameFunc != nil {
-		return t.spanNameFunc(stmt)
-	}
-
+func defaultSpanNameFunc(stmt string) string {
 	stmt = strings.TrimSpace(stmt)
 	end := strings.IndexFunc(stmt, unicode.IsSpace)
 	if end < 0 && len(stmt) > 0 {